+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 - Now you can access the params used in a method directly from the callback function (`on('method', (params) => ...)`). You can also add types to the params in the `on` method (`on('method', (params: string) => ...)`). Note that you can't pass interfaces in this case, just types.
-- There is a new method `requestParams` that returns the parameters of the current method. The method `params` is now deprecated, as it would return just the first parameter as a string. However, it is highly recommended to get the current params via the `params` argument in the callback function, as mentioned above.
+- There is a new method `requestParams` that returns the array of parameters of the current method. The method `params` is now deprecated, as it would return just the first parameter as a string. However, it is highly recommended to get the current params via the `params` argument in the callback function, as mentioned above.
 - When a method that does not exist is called, an error will be thrown, indicating the misspelled method.
 
 ## 2.1.0